handler_audio.cc: Don't rely on SDL_mixer using libmikmod
As on different platforms it may use libmodplug.
Instead, use mikmod directly for music and catch output with
Mix_HookMusic().

This is pretty hacky as 'no sound' driver is used with
direct calling to VC_WriteBytes() to fill the buffer.

MikMod_RegisterAllLoaders() fail for some platforms so use
MikMod_RegisterLoader() separately.

Register .xm loader as well as .mod because area1-game2.mod is .xm
in disguise.